Remove combobox item template selector and add data template instead. Add BaseValueToAttribute converter, rename EnumDescriptionConverter and remove IsEnum check.
