+++ Controls/PropertiesEditor/PropertyEditorUC.xaml
@@ -98,19 +98,15 @@
             </DataTemplate>
 
             <!-- the 4 choice modes -->
-            <!-- 1. choice default mode, if other choice templates need to show description instead of value, apply itemtemplateselector from here to the other templates-->
+            <!-- 1. choice default mode-->
             <DataTemplate DataType="{x:Type pi:ChoicePropertyProxy}" x:Key="choiceProxy">
                 <ComboBox DisplayMemberPath="{Binding DisplayMemberPath}" HorizontalAlignment="Stretch" x:Name="PART_EDIT"
                     ItemsSource="{Binding Values}" SelectedItem="{Binding Value}">
-                    <ComboBox.ItemTemplateSelector>
-                        <loc:ComboBoxItemTemplateSelector>
-                            <loc:ComboBoxItemTemplateSelector.DescriptionDataTemplate>
-                                <DataTemplate>
-                                    <TextBlock Text="{Binding Converter={locconv:EnumDescriptionConverter}}" />
-                                </DataTemplate>
-                            </loc:ComboBoxItemTemplateSelector.DescriptionDataTemplate>
-                        </loc:ComboBoxItemTemplateSelector>
-                    </ComboBox.ItemTemplateSelector>
+                    <ComboBox.ItemTemplate>
+                        <DataTemplate>
+                            <TextBlock Text="{Binding Converter={locconv:ValueToDescriptionConverter}}" />
+                        </DataTemplate>
+                    </ComboBox.ItemTemplate>
                 </ComboBox>
             </DataTemplate>
             <!-- choice async mode-->
@@ -133,7 +129,7 @@
             <!-- choice editable mode -->
             <DataTemplate DataType="{x:Type pi:ChoicePropertyProxy}" x:Key="choiceEditableProxy">
                 <ComboBox DisplayMemberPath="{Binding DisplayMemberPath}" HorizontalAlignment="Stretch" x:Name="PART_EDIT"
-                    ItemsSource="{Binding Values}" IsEditable="True" Text="{Binding Value}"/>
+                    ItemsSource="{Binding Values}" IsEditable="True" Text="{Binding Value}" />
             </DataTemplate>
             <!-- choice editable async mode -->
             <DataTemplate DataType="{x:Type pi:ChoicePropertyProxy}" x:Key="choiceEditableAsyncProxy">
